Extract response helpers to module
Allow developers to use Stripe test data when stubbing responses
locally. This should help mitigate the risk of API drift with stubbed
responses.

* Move response helpers to their own file
